backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0

# Telegram Bot dependencies
python-telegram-bot==20.3
APScheduler==3.10.4
python-dotenv==1.0.0
# psycopg2-binary==2.9.9  # Only needed for PostgreSQL production

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
pytest-cov==4.1.0
requests-mock==1.11.0
freezegun==1.2.2